move the digitizer faux py files from bapsflib._hdf.maps.digitizers.tests -> bapsflib.faux.digitizers
