+++ bapsflib/_hdf/maps/digitizers/tests/__init__.py
@@ -1,3 +1,1 @@
-from bapsflib._hdf.maps.digitizers.tests.fauxlecroy180e import FauxLeCroy180E
-from bapsflib._hdf.maps.digitizers.tests.fauxsis3301 import FauxSIS3301
-from bapsflib._hdf.maps.digitizers.tests.fauxsiscrate import FauxSISCrate
+
