+++ bapsflib/_hdf/utils/tests/test_hdfreaddata.py
@@ -17,7 +17,6 @@ import astropy.units as u
 import h5py
 import numpy as np
 import os
-import unittest as ut
 
 from typing import TYPE_CHECKING
 from unittest import mock
@@ -38,8 +37,8 @@ from bapsflib.utils.warnings import BaPSFWarning, HDFMappingWarning
 
 if TYPE_CHECKING:
     # This is done for typing purposes only.  A full import is not needed.
-    from bapsflib._hdf.maps.digitizers.tests.fauxlecroy180e import FauxLeCroy180E
-    from bapsflib._hdf.maps.digitizers.tests.fauxsis3301 import FauxSIS3301
+    from bapsflib.faux.digitizers.fauxlecroy180e import FauxLeCroy180E
+    from bapsflib.faux.digitizers.fauxsis3301 import FauxSIS3301
 
 
 class TestHDFReadData(TestBase):
