+++ bapsflib/faux/digitizers/fauxlecroy180e.py
