+++ bapsflib/faux/digitizers/fauxsis3301.py
