+++ bapsflib/faux/digitizers/fauxsiscrate.py
